Feature/backend

+++ sprintly/src/main/java/com/sprintly/sprintly/controller/OrganizationController.java
@@ -1,14 +1,14 @@
 package com.sprintly.sprintly.controller;
 
 import com.sprintly.sprintly.entity.Organization;
-import com.sprintly.sprintly.entity.User;
 import com.sprintly.sprintly.entity.UserOrganizationRole;
+import com.sprintly.sprintly.exception.custom.CustomException;
 import com.sprintly.sprintly.model.Organization.OrganizationDeleteDto;
 import com.sprintly.sprintly.model.Organization.OrganizationDto;
-import com.sprintly.sprintly.model.auth.RegisterUserDto;
 import com.sprintly.sprintly.service.organization.OrganizationService;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -22,18 +22,28 @@ public class OrganizationController {
     private OrganizationService organizationService;
 
     @PostMapping("/create")
-    public Organization registerUser(@RequestBody OrganizationDto organizationDto) {
+    public Organization createOrganization(@RequestBody OrganizationDto organizationDto) {
         return organizationService.create(organizationDto);
     }
 
     @PostMapping("/all")
-    public List<UserOrganizationRole> getAllUserOrganizationRole(@RequestParam String emailID){
+    public List<UserOrganizationRole> getAllUserOrganizationRole(@RequestParam String emailID) {
         return organizationService.getAllUserOrganizationRole(emailID);
     }
 
-    @DeleteMapping("/delete/name")
-    public void deleteOrganization(@RequestBody OrganizationDeleteDto organizationDeleteDto){
-        organizationService.deleteOrganization(organizationDeleteDto);
+    @DeleteMapping("/delete")
+    public String deleteOrganization(
+            @RequestBody OrganizationDeleteDto organizationDeleteDto,
+            Authentication authentication) {
+
+        String jwtEmail = authentication.getName(); // Typically, the username/email
+        String dtoEmail = organizationDeleteDto.getEmailID();
+
+        if (!jwtEmail.equals(dtoEmail)) {
+            throw new CustomException("Email in JWT does not match email in request.");
+        }
+
+        return organizationService.deleteOrganization(organizationDeleteDto);
     }
 
 }
